to orig dtype in emb

+++ owl_wms/trainers/world_trainer.py
@@ -219,7 +219,8 @@ class WorldTrainer(BaseTrainer):
         batch = {k: v.cuda() if isinstance(v, torch.Tensor) else v for k, v in batch.items()}
         if "rgb" in batch:
             assert "x" not in batch, "passed rgb to convert, but already have batch item `x` (latents)"
-            batch["x"] = self.encoder_decoder.encode(batch.pop("rgb")).bfloat16()
+            batch["x"] = self.encoder_decoder.encode(batch.pop("rgb"))
+        batch["x"] = batch["x"].bfloat16()
         if "prompt" in batch:
             assert "prompt_emb" not in batch, "passed prompt to convert, but already have batch item `prompt_emb`"
             batch["prompt_emb"] = self.prompt_encoder(batch.pop("prompt")).to(batch["x"].dtype)
